refactor: split event action into separate files
