+++ code/handlers/event_msg_action.go
@@ -1,0 +1,38 @@
+package handlers
+
+import (
+	"fmt"
+	"start-feishubot/services/openai"
+)
+
+type MessageAction struct { /*消息*/
+}
+
+func (*MessageAction) Execute(a *ActionInfo) bool {
+	msg := a.handler.sessionCache.GetMsg(*a.info.sessionId)
+	msg = append(msg, openai.Messages{
+		Role: "user", Content: a.info.qParsed,
+	})
+	completions, err := a.handler.gpt.Completions(msg)
+	if err != nil {
+		replyMsg(*a.ctx, fmt.Sprintf(
+			"🤖️：消息机器人摆烂了，请稍后再试～\n错误信息: %v", err), a.info.msgId)
+		return false
+	}
+	msg = append(msg, completions)
+	a.handler.sessionCache.SetMsg(*a.info.sessionId, msg)
+	//if new topic
+	if len(msg) == 2 {
+		//fmt.Println("new topic", msg[1].Content)
+		sendNewTopicCard(*a.ctx, a.info.sessionId, a.info.msgId,
+			completions.Content)
+		return false
+	}
+	err = replyMsg(*a.ctx, completions.Content, a.info.msgId)
+	if err != nil {
+		replyMsg(*a.ctx, fmt.Sprintf(
+			"🤖️：消息机器人摆烂了，请稍后再试～\n错误信息: %v", err), a.info.msgId)
+		return false
+	}
+	return true
+}
